Remove npermute, there is no general implementation for this function.
There is not an algorithm for the destructive permutation of a vector or
matrix. This applies to the destructive transpose of a permutation matrix
as well. Since the permutation matrix contains cycles with fixed points,
naive destructive permutation algorithms endlessly repeat and do not
permute rows/columns outside of the first cycle encountered. This is the
same reason that a permutation matrix cannot be destructively tranposed.